refactor: move deploy reducer to useDeploy hook
